Crate drone_cortex_m

ARM® Cortex®-M platform crate for Drone, an Embedded Operating System.

Supported Cores

Architecture	Core name	Rust target	cortex_m_core config flag

ARMv7-M	ARM® Cortex®-M3 r0p0	thumbv7m-none-eabi	cortex_m3_r0p0

ARMv7-M	ARM® Cortex®-M3 r1p0	thumbv7m-none-eabi	cortex_m3_r1p0

ARMv7-M	ARM® Cortex®-M3 r1p1	thumbv7m-none-eabi	cortex_m3_r1p1

ARMv7-M	ARM® Cortex®-M3 r2p0	thumbv7m-none-eabi	cortex_m3_r2p0

ARMv7-M	ARM® Cortex®-M3 r2p1	thumbv7m-none-eabi	cortex_m3_r2p1

ARMv7E-M	ARM® Cortex®-M4 r0p0	thumbv7em-none-eabi	cortex_m4_r0p0

ARMv7E-M	ARM® Cortex®-M4 r0p1	thumbv7em-none-eabi	cortex_m4_r0p1

ARMv7E-M	ARM® Cortex®-M4F r0p0	thumbv7em-none-eabihf	cortex_m4f_r0p0

ARMv7E-M	ARM® Cortex®-M4F r0p1	thumbv7em-none-eabihf	cortex_m4f_r0p1

Rust target triple and cortex_m_core config flag should be set at the application level according to this table.

Documentation

Drone Book

API documentation

Usage

Place the following to the Cargo.toml:

[dependencies]
drone-cortex-m = { version = "0.11.0", features = [...] }

Modules

drv
Drivers for core ARM Cortex-M peripherals.

fib
The Fibers module.

itm
The Instrumentation Trace Macrocell.

map
Core ARM Cortex-M register and exception mappings.

prelude
The Drone Cortex-M Prelude.

proc_loop
This module provides interface to wrap a stackful synchronous code into an asynchronous command loop.

processor
Common utility functions for working with ARM Cortex-M processors.

reg
The Memory-Mapped Registers module.

reg
The Memory-Mapped Registers module.

sv
The Supervisor module.

thr
The Threads module.

thr
The Threads module.

Macros

dbg
Prints and returns the value of a given expression for quick and dirty debugging.

eprint
Prints to the ITM port #1, if a debug probe is connected.

eprintln
Prints to the ITM port #1, with a newline, if a debug probe is connected.

print
Prints to the ITM port #0, if a debug probe is connected.

println
Prints to the ITM port #0, with a newline, if a debug probe is connected.

reg

sv
Defines the supervisor type.

thr