Module drone_cortex_m::itm

The Instrumentation Trace Macrocell.

This module provides interface to transmit log data through the SWO pin.

ITM ports #0 and #1 are reserved for STDOUT and STDERR respectively.

Structs

Port
ITM stimulus port handler.

Constants

HEAP_TRACE_KEY
XOR pattern for heap trace output.

HEAP_TRACE_PORT
Port number of the heap trace stream.

STDERR_PORT
Port number of the standard error stream.

STDOUT_PORT
Port number of the standard output stream.

Functions

flush
Blocks until all pending packets are transmitted.

is_enabled
Returns true if a debug probe is connected and listening to the ITM output.

is_port_enabled
Returns true if a debug probe is connected and listening to the ITM port port output.

sync
Generates an ITM synchronization packet.

trace_alloc
Logs an allocation to the ITM port #31.

trace_dealloc
Logs a deallocation to the ITM port #31.

trace_grow_in_place
Logs growing in place to the ITM port #31.

trace_shrink_in_place
Logs shrinking in place to the ITM port #31.

update_prescaler
Updates the SWO prescaler register.

write_fmt
Writes args to the stimulus port number address.

write_str
Writes string to the stimulus port number address.