Struct drone_cortex_m::map::reg::scb::aircr::Hold

pub struct Hold<'a, _T: RegTag> { /* fields omitted */ }

Application interrupt and reset control register.

Methods

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn vectkey(&self) -> u32

Vector Key.

Register writes must write 0x05FA to this field, otherwise the write is ignored.

On reads, returns 0xFA05 .

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn write_vectkey(&mut self, bits: u32) -> &mut Self

Vector Key.

Register writes must write 0x05FA to this field, otherwise the write is ignored.

On reads, returns 0xFA05 .

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn endianess(&self) -> bool

Data endianness bit.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn prigroup(&self) -> u32

Interrupt priority grouping field.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn write_prigroup(&mut self, bits: u32) -> &mut Self

Interrupt priority grouping field.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn set_sysresetreq(&mut self) -> &mut Self

System reset request.

pub fn clear_sysresetreq(&mut self) -> &mut Self

System reset request.

pub fn toggle_sysresetreq(&mut self) -> &mut Self

System reset request.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn set_vectclractive(&mut self) -> &mut Self

Clears all active state information for exceptions.

pub fn clear_vectclractive(&mut self) -> &mut Self

Clears all active state information for exceptions.

pub fn toggle_vectclractive(&mut self) -> &mut Self

Clears all active state information for exceptions.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn set_vectreset(&mut self) -> &mut Self

Resets the processor (except debug logic), but this will not reset circuits outside the processor.

pub fn clear_vectreset(&mut self) -> &mut Self

Resets the processor (except debug logic), but this will not reset circuits outside the processor.

pub fn toggle_vectreset(&mut self) -> &mut Self

Resets the processor (except debug logic), but this will not reset circuits outside the processor.