Struct drone_cortex_m::map::reg::scb::demcr::Hold

pub struct Hold<'a, _T: RegTag> { /* fields omitted */ }

Debug Exception and Monitor Control Register.

Methods

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn trcena(&self) -> bool

Global enable for all DWT and ITM features.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn set_trcena(&mut self) -> &mut Self

Global enable for all DWT and ITM features.

pub fn clear_trcena(&mut self) -> &mut Self

Global enable for all DWT and ITM features.

pub fn toggle_trcena(&mut self) -> &mut Self

Global enable for all DWT and ITM features.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn mon_req(&self) -> bool

DebugMonitor semaphore bit.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn set_mon_req(&mut self) -> &mut Self

DebugMonitor semaphore bit.

pub fn clear_mon_req(&mut self) -> &mut Self

DebugMonitor semaphore bit.

pub fn toggle_mon_req(&mut self) -> &mut Self

DebugMonitor semaphore bit.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn mon_step(&self) -> bool

Setting this bit to 1 makes the step request pending.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn set_mon_step(&mut self) -> &mut Self

Setting this bit to 1 makes the step request pending.

pub fn clear_mon_step(&mut self) -> &mut Self

Setting this bit to 1 makes the step request pending.

pub fn toggle_mon_step(&mut self) -> &mut Self

Setting this bit to 1 makes the step request pending.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn mon_pend(&self) -> bool

Sets or clears the pending state of the DebugMonitor exception.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn set_mon_pend(&mut self) -> &mut Self

Sets or clears the pending state of the DebugMonitor exception.

pub fn clear_mon_pend(&mut self) -> &mut Self

Sets or clears the pending state of the DebugMonitor exception.

pub fn toggle_mon_pend(&mut self) -> &mut Self

Sets or clears the pending state of the DebugMonitor exception.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn mon_en(&self) -> bool

Enable the DebugMonitor exception.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn set_mon_en(&mut self) -> &mut Self

Enable the DebugMonitor exception.

pub fn clear_mon_en(&mut self) -> &mut Self

Enable the DebugMonitor exception.

pub fn toggle_mon_en(&mut self) -> &mut Self

Enable the DebugMonitor exception.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn vc_harderr(&self) -> bool

Enable halting debug trap on a HardFault exception.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn set_vc_harderr(&mut self) -> &mut Self

Enable halting debug trap on a HardFault exception.

pub fn clear_vc_harderr(&mut self) -> &mut Self

Enable halting debug trap on a HardFault exception.

pub fn toggle_vc_harderr(&mut self) -> &mut Self

Enable halting debug trap on a HardFault exception.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn vc_interr(&self) -> bool

Enable halting debug trap on a fault occurring during exception entry or exception return.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn set_vc_interr(&mut self) -> &mut Self

Enable halting debug trap on a fault occurring during exception entry or exception return.

pub fn clear_vc_interr(&mut self) -> &mut Self

Enable halting debug trap on a fault occurring during exception entry or exception return.

pub fn toggle_vc_interr(&mut self) -> &mut Self

Enable halting debug trap on a fault occurring during exception entry or exception return.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn vc_buserr(&self) -> bool

Enable halting debug trap on a BusFault exception.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn set_vc_buserr(&mut self) -> &mut Self

Enable halting debug trap on a BusFault exception.

pub fn clear_vc_buserr(&mut self) -> &mut Self

Enable halting debug trap on a BusFault exception.

pub fn toggle_vc_buserr(&mut self) -> &mut Self

Enable halting debug trap on a BusFault exception.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn vc_staterr(&self) -> bool

Enable halting debug trap on a UsageFault exception caused by a state information error, for example an Undefined Instruction exception.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn set_vc_staterr(&mut self) -> &mut Self

Enable halting debug trap on a UsageFault exception caused by a state information error, for example an Undefined Instruction exception.

pub fn clear_vc_staterr(&mut self) -> &mut Self

Enable halting debug trap on a UsageFault exception caused by a state information error, for example an Undefined Instruction exception.

pub fn toggle_vc_staterr(&mut self) -> &mut Self

Enable halting debug trap on a UsageFault exception caused by a state information error, for example an Undefined Instruction exception.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn vc_chkerr(&self) -> bool

Enable halting debug trap on a UsageFault exception caused by a checking error, for example an alignment check error.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn set_vc_chkerr(&mut self) -> &mut Self

Enable halting debug trap on a UsageFault exception caused by a checking error, for example an alignment check error.

pub fn clear_vc_chkerr(&mut self) -> &mut Self

Enable halting debug trap on a UsageFault exception caused by a checking error, for example an alignment check error.

pub fn toggle_vc_chkerr(&mut self) -> &mut Self

Enable halting debug trap on a UsageFault exception caused by a checking error, for example an alignment check error.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn vc_nocperr(&self) -> bool

Enable halting debug trap on a UsageFault caused by an access to a Coprocessor.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn set_vc_nocperr(&mut self) -> &mut Self

Enable halting debug trap on a UsageFault caused by an access to a Coprocessor.

pub fn clear_vc_nocperr(&mut self) -> &mut Self

Enable halting debug trap on a UsageFault caused by an access to a Coprocessor.

pub fn toggle_vc_nocperr(&mut self) -> &mut Self

Enable halting debug trap on a UsageFault caused by an access to a Coprocessor.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn vc_mmerr(&self) -> bool

Enable halting debug trap on a MemManage exception.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn set_vc_mmerr(&mut self) -> &mut Self

Enable halting debug trap on a MemManage exception.

pub fn clear_vc_mmerr(&mut self) -> &mut Self

Enable halting debug trap on a MemManage exception.

pub fn toggle_vc_mmerr(&mut self) -> &mut Self

Enable halting debug trap on a MemManage exception.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn vc_corereset(&self) -> bool

Enable Reset Vector Catch. This causes a Local reset to halt a running system.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn set_vc_corereset(&mut self) -> &mut Self

Enable Reset Vector Catch. This causes a Local reset to halt a running system.

pub fn clear_vc_corereset(&mut self) -> &mut Self

Enable Reset Vector Catch. This causes a Local reset to halt a running system.

pub fn toggle_vc_corereset(&mut self) -> &mut Self

Enable Reset Vector Catch. This causes a Local reset to halt a running system.