Struct drone_cortex_m::map::reg::scb::mmfsr::Hold

pub struct Hold<'a, _T: RegTag> { /* fields omitted */ }

MemManage Status Register.

Methods

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn mmarvalid(&self) -> bool

MMFAR has valid contents.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn set_mmarvalid(&mut self) -> &mut Self

MMFAR has valid contents.

pub fn clear_mmarvalid(&mut self) -> &mut Self

MMFAR has valid contents.

pub fn toggle_mmarvalid(&mut self) -> &mut Self

MMFAR has valid contents.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn mlsperr(&self) -> bool

A MemManage fault occurred during FP lazy state preservation.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn set_mlsperr(&mut self) -> &mut Self

A MemManage fault occurred during FP lazy state preservation.

pub fn clear_mlsperr(&mut self) -> &mut Self

A MemManage fault occurred during FP lazy state preservation.

pub fn toggle_mlsperr(&mut self) -> &mut Self

A MemManage fault occurred during FP lazy state preservation.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn mstkerr(&self) -> bool

A derived MemManage fault occurred on exception entry.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn set_mstkerr(&mut self) -> &mut Self

A derived MemManage fault occurred on exception entry.

pub fn clear_mstkerr(&mut self) -> &mut Self

A derived MemManage fault occurred on exception entry.

pub fn toggle_mstkerr(&mut self) -> &mut Self

A derived MemManage fault occurred on exception entry.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn munstkerr(&self) -> bool

A derived MemManage fault occurred on exception return.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn set_munstkerr(&mut self) -> &mut Self

A derived MemManage fault occurred on exception return.

pub fn clear_munstkerr(&mut self) -> &mut Self

A derived MemManage fault occurred on exception return.

pub fn toggle_munstkerr(&mut self) -> &mut Self

A derived MemManage fault occurred on exception return.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn daccviol(&self) -> bool

Data access violation. The MMFAR shows the data address that the load or store tried to access.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn set_daccviol(&mut self) -> &mut Self

Data access violation. The MMFAR shows the data address that the load or store tried to access.

pub fn clear_daccviol(&mut self) -> &mut Self

Data access violation. The MMFAR shows the data address that the load or store tried to access.

pub fn toggle_daccviol(&mut self) -> &mut Self

Data access violation. The MMFAR shows the data address that the load or store tried to access.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn iaccviol(&self) -> bool

MPU or Execute Never (XN) default memory map access violation on an instruction fetch has occurred. The fault is signalled only if the instruction is issued.

impl<'a, _T: RegTag> Hold<'a, _T>

pub fn set_iaccviol(&mut self) -> &mut Self

MPU or Execute Never (XN) default memory map access violation on an instruction fetch has occurred. The fault is signalled only if the instruction is issued.

pub fn clear_iaccviol(&mut self) -> &mut Self

MPU or Execute Never (XN) default memory map access violation on an instruction fetch has occurred. The fault is signalled only if the instruction is issued.

pub fn toggle_iaccviol(&mut self) -> &mut Self

MPU or Execute Never (XN) default memory map access violation on an instruction fetch has occurred. The fault is signalled only if the instruction is issued.