Module drone_cortex_m::prelude

The Drone Cortex-M Prelude.

NOTE This module documentation should be viewed as a continuation of the drone_core documentation.

By default Rust automatically injects libcore prelude imports into every module. To inject the Drone prelude instead, place the following code to the src/lib.rs:

#![feature(prelude_import)]

#[prelude_import]
#[allow(unused_imports)]
use drone_cortex_m::prelude::*;

Re-exports

pub use drone_core::prelude::*;

pub use crate::dbg;

pub use crate::eprint;

pub use crate::eprintln;

pub use crate::print;

pub use crate::println;