Function drone_cortex_m::processor::self_reset

pub fn self_reset() -> !

Requests system reset.

Generates a system reset request to the microcontroller's system reset control logic. Because the system reset control logic is not a part of the processor design, the exact timing of the reset is device-specific.

The debug logic is not affected.