Trait drone_cortex_m::reg::field::RRRegFieldBitBand

pub trait RRRegFieldBitBand<T: RegTag> where
    Self: RRRegFieldBit<T>,
    Self::Reg: RegBitBand<T> + RReg<T>, {
    fn read_bit_band(&self) -> bool;

    fn to_bit_band_ptr(&self) -> *const usize;
}

Readable single-bit field of readable register located in the peripheral bit-band region.

Required methods

fn read_bit_band(&self) -> bool

Reads the value of this bit through the peripheral bit-band region alias.

fn to_bit_band_ptr(&self) -> *const usize

Returns a raw pointer to the bit-band alias address of this field.

See also WWRegFieldBitBand::to_bit_band_mut_ptr.

Loading content...