Trait drone_cortex_m::reg::marker::WoWoRegFieldBit

pub trait WoWoRegFieldBit<T>: RegFieldBit<T> + WoWRegField<T> where
    T: RegTag,
    Self::Reg: WoReg<T>, {
    fn set_bit(&self);

    fn clear_bit(&self);

    fn toggle_bit(&self);
}

Write-only single-bit field of write-only register.

Required methods

fn set_bit(&self)

Writes the reset value with the bit set into the register memory.

fn clear_bit(&self)

Writes the reset value with the bit cleared into the register memory.

fn toggle_bit(&self)

Writes the reset value with the bit toggled into the register memory.

Loading content...