Trait drone_cortex_m::thr::prelude::_

pub trait _: ThrToken {
    fn add_future<F, Y, T>(self, fib: F) -> FiberFuture<T>
    where
        F: Fiber<Input = (), Yield = Y, Return = T> + Send + 'static,
        T: Send + 'static,
        Y: YieldNone,
    { ... }
}

Extends ThrToken types with add_future method.

Provided methods

fn add_future<F, Y, T>(self, fib: F) -> FiberFuture<T> where
F: Fiber<Input = (), Yield = Y, Return = T> + Send + 'static,
T: Send + 'static,
Y: YieldNone,

Adds the fiber fib to the fiber chain and returns a future, which resolves on completion of the fiber.

Loading content...