Trait drone_cortex_m::thr::ThrNvic

pub trait ThrNvic: IntToken {
    fn enable_batch<F>(&self, f: F)
    where
        F: FnOnce(&mut NvicIser<Self::Bundle>),
    { ... }

    fn enable(&self, bundle: &mut NvicIser<Self::Bundle>) { ... }

    fn enable_int(&self) { ... }

    fn enabled(&self) -> NvicIser<Self::Bundle> { ... }

    fn is_enabled(&self, bundle: &NvicIser<Self::Bundle>) -> bool { ... }

    fn is_int_enabled(&self) -> bool { ... }

    fn disable_batch<F>(&self, f: F)
    where
        F: FnOnce(&mut NvicIcer<Self::Bundle>),
    { ... }

    fn disable(&self, bundle: &mut NvicIcer<Self::Bundle>) { ... }

    fn disable_int(&self) { ... }

    fn set_pending_batch<F>(&self, f: F)
    where
        F: FnOnce(&mut NvicIspr<Self::Bundle>),
    { ... }

    fn set_pending(&self, bundle: &mut NvicIspr<Self::Bundle>) { ... }

    fn set_pending_int(&self) { ... }

    fn clear_pending_batch<F>(&self, f: F)
    where
        F: FnOnce(&mut NvicIcpr<Self::Bundle>),
    { ... }

    fn clear_pending(&self, bundle: &mut NvicIcpr<Self::Bundle>) { ... }

    fn clear_pending_int(&self) { ... }

    fn pending(&self) -> NvicIcpr<Self::Bundle> { ... }

    fn is_pending(&self, bundle: &NvicIcpr<Self::Bundle>) -> bool { ... }

    fn is_int_pending(&self) -> bool { ... }

    fn active(&self) -> NvicIabr<Self::Bundle> { ... }

    fn is_active(&self, bundle: &NvicIabr<Self::Bundle>) -> bool { ... }

    fn is_int_active(&self) -> bool { ... }

    fn priority(&self) -> u8 { ... }

    fn set_priority(&self, priority: u8) { ... }
}

NVIC methods for interrupt tokens.

Provided methods

fn enable_batch<F>(&self, f: F) where
F: FnOnce(&mut NvicIser<Self::Bundle>),

Enables multiple interrupts within the NVIC bundle.

fn enable(&self, bundle: &mut NvicIser<Self::Bundle>)

Enables the interrupt.

fn enable_int(&self)

Enables the interrupt within the NVIC bundle.

fn enabled(&self) -> NvicIser<Self::Bundle>

Returns the NVIC bundle of enabled states.

fn is_enabled(&self, bundle: &NvicIser<Self::Bundle>) -> bool

Returns true if the interrupt is enabled.

fn is_int_enabled(&self) -> bool

Returns true if the interrupt is enabled within the NVIC bundle.

fn disable_batch<F>(&self, f: F) where
F: FnOnce(&mut NvicIcer<Self::Bundle>),

Disables multiple interrupts within the NVIC bundle.

fn disable(&self, bundle: &mut NvicIcer<Self::Bundle>)

Disables the interrupt.

fn disable_int(&self)

Disables the interrupt within the NVIC bundle.

fn set_pending_batch<F>(&self, f: F) where
F: FnOnce(&mut NvicIspr<Self::Bundle>),

Sets multiple interrupts pending within the NVIC bundle.

fn set_pending(&self, bundle: &mut NvicIspr<Self::Bundle>)

Sets the interrupt pending.

fn set_pending_int(&self)

Sets the interrupt pending within the NVIC bundle.

fn clear_pending_batch<F>(&self, f: F) where
F: FnOnce(&mut NvicIcpr<Self::Bundle>),

Clears multiple interrupts pending state within the NVIC bundle.

fn clear_pending(&self, bundle: &mut NvicIcpr<Self::Bundle>)

Clears the interrupt pending state.

fn clear_pending_int(&self)

Clears the interrupt pending state within the NVIC bundle.

fn pending(&self) -> NvicIcpr<Self::Bundle>

Returns the NVIC bundle of pending states.

fn is_pending(&self, bundle: &NvicIcpr<Self::Bundle>) -> bool

Returns true if the interrupt is pending.

fn is_int_pending(&self) -> bool

Returns true if the interrupt is pending within the NVIC bundle.

fn active(&self) -> NvicIabr<Self::Bundle>

Returns the NVIC bundle of active states.

fn is_active(&self, bundle: &NvicIabr<Self::Bundle>) -> bool

Returns true if the interrupt is active.

fn is_int_active(&self) -> bool

Returns true if the interrupt is active within the NVIC bundle.

fn priority(&self) -> u8

Reads the priority of the interrupt.

fn set_priority(&self, priority: u8)

Writes the priority of the interrupt.

Loading content...