Module drone_cortex_m::map::reg::fpu::fpccr

Floating-point context control register.

Structs

Aspen
When this bit is set to 1, execution of a floating-point instruction sets the CONTROL.FPCA bit to 1.

Bfrdy
Indicates whether the software executing when the processor allocated the FP stack frame was able to set the BusFault exception to pending.

Hfrdy
Indicates whether the software executing when the processor allocated the FP stack frame was able to set the HardFault exception to pending.

Hold
Floating-point context control register.

Lspact
Indicates whether Lazy preservation of the FP state is active.

Lspen
Enables lazy context save of FP state.

Mmrdy
Indicates whether the software executing when the processor allocated the FP stack frame was able to set the MemManage exception to pending.

Monrdy
Indicates whether the software executing when the processor allocated the FP stack frame was able to set the DebugMonitor exception to pending.

Reg
Floating-point context control register.

Thread
Indicates the processor mode when it allocated the FP stack frame.

User
Indicates the privilege level of the software executing when the processor allocated the FP stack frame.

Val
Floating-point context control register.